requirements.txt
#---- Backend ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
pydantic>=2.7.1

# --- ML / Vision ---
torch>=2.9.0
torchvision>=0.20.0
transformers>=4.41.2
Pillow>=10.3.0
accelerate>=0.30.0

#---- Frontend --
streamlit>=1.35.0
plotly>=5.22.0
requests>=2.32.2
